tests(core): Add a new test folder to the app and refactor old tests
